+++ packages/sync-rules/src/utils.ts
@@ -1,21 +1,20 @@
 import { JSONBig, JsonContainer, Replacer, stringifyRaw } from '@powersync/service-jsonbig';
 import { SelectFromStatement, Statement } from 'pgsql-ast-parser';
+import { CompatibilityContext } from './compatibility.js';
+import { SyncRuleProcessingError as SyncRulesProcessingError } from './errors.js';
 import { SQLITE_FALSE, SQLITE_TRUE } from './sql_support.js';
 import {
+  BucketIdTransformer,
   DatabaseInputRow,
   DatabaseInputValue,
-  SqliteInputValue,
   SqliteInputRow,
+  SqliteInputValue,
   SqliteJsonRow,
   SqliteJsonValue,
   SqliteRow,
-  SqliteValue,
-  BucketIdTransformer,
-  ToastableSqliteRow
+  SqliteValue
 } from './types.js';
-import { SyncRuleProcessingError as SyncRulesProcessingError } from './errors.js';
 import { CustomArray, CustomObject, CustomSqliteValue } from './types/custom_sqlite_value.js';
-import { CompatibilityContext } from './compatibility.js';
 
 export function isSelectStatement(q: Statement): q is SelectFromStatement {
   return q.type == 'select';