+++ modules/module-core/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @powersync/service-module-core
 
+## 0.2.12
+
+### Patch Changes
+
+- Updated dependencies [fff0024]
+  - @powersync/service-core@1.16.3
+  - @powersync/lib-services-framework@0.7.10
+  - @powersync/service-rsocket-router@0.2.7
+
 ## 0.2.11
 
 ### Patch Changes